Omit checking for new functions during 20h, see http://zsh.sourceforge.net/Doc/Release/Completion-System.html

I get a speed bump from 0.06s to 0.04s on zsh startup.

Don't know how this interacts with zimfw compile (we compile zcompdump). Code was taken from prezto/omz