@ericbn 2019-12-10
I like the idea of having some mechanism to update the dumpfile, as indeed we never allow it to be updated in the current code, as we always skip the check with `-C`. But we have something in our favor, that is the install/update actions in zimfw. That's when things can change for the completion: when a module is installed or updated. So maybe we can add some logic there instead.