@rjcoelho 2019-12-10
> I like the idea of having some mechanism to update the dumpfile, as indeed we never allow it to be updated in the current code, as we always skip the check with `-C`. But we have something in our favor, that is the install/update actions in zimfw. That's when things can change for the completion: when a module is installed or updated. So maybe we can add some logic there instead.
I also thought about it. 

Correct me if i'm wrong but currently if we add/remove completions we should zimfw clean (rm zdumpfile) the file, then do a login (which generates a new zdumpfile since we call it with -C) and then zimfw compile (to build .zwc) ? 

This change basically automates it by "forcing" to generate a new zdumpfile every 20h. Unlike what I say, it isn't a performance improvement. Every 20h we force a new zdumpfile (by calling w/o -C) which we should then zcompile it (we should probably zcompile it here in this PR). What do you think ?