@ericbn 2019-12-11
Yes, currently the only way to force compinit to dump a new configuration is to run `zimfw clean-dumpfile`.

As explained in the [compinit documentation](zsh.sourceforge.net/Doc/Release/Completion-System.html#Use-of-compinit)
> If the number of completion files changes, <tt>compinit</tt> will recognise this and produce a new dump file.  However, if the name of a function or the arguments in the first line of a <tt>#compdef</tt> function (as described below) change, it is easiest to delete the dump file by hand so that <tt>compinit</tt> will re-create it the next time it is run.  The check performed to see if there are new functions can be omitted by giving the option <tt>-C</tt>.  In this case the dump file will only be created if there isn’t one already.

I see a couple issues with the proposed solution:
1. The 20 hours expiration is arbitrary and will impact the startup performance every 20 hours. Startup time is precious to us. Depending on the user configuration and how often he/she updates it, the completion would require effective regeneration every 20 hours, 20 days, 20 weeks, 20 months...  So we don't want to penalize all users every 20 hours.
2. The way compinit recognizes that the configuration needs to be updated is too simple, and only covers the scenario when the number of completion files changes. All other scenarios still require the dumpfile to be manually deleted.

But you raised a very good point, and I already have an idea on how to cover all scenarios during the usage of the zimfw tool, which also has the advantage of not being executed during start up. Will work on it soon.