@ericbn 2019-12-11
Benchmark with the default zimfw installation (from the develop branch as the time of writing this):

* with `-C` (which is how it currently works):
    ```
    time                 34.34 ms   (34.17 ms .. 34.59 ms)
                         1.000 R²   (0.999 R² .. 1.000 R²)
    mean                 34.67 ms   (34.48 ms .. 34.86 ms)
    std dev              405.1 μs   (334.0 μs .. 524.8 μs)
    ```

* without `-C`:
    ```
    time                 66.78 ms   (66.16 ms .. 68.10 ms)
                         1.000 R²   (0.999 R² .. 1.000 R²)
    mean                 66.47 ms   (66.13 ms .. 66.91 ms)
    std dev              640.3 μs   (420.8 μs .. 985.6 μs)
    ```

So impact is significant! (And benefit is debatable, as I pointed out in the previous comment.)