@rjcoelho 2019-12-11
So we should always call w/o -C. 

I want do be able to install/delete/update packages on my machine without worrying if it needs to dump a new compinit. That is the problem I am trying to solve here.

The 20h is a "daily" interval. Should probably be 24h. So I pay a 30ms penalty on the first daily login. It's acceptable imo. 

Every day I will call once without -C and then zcompile it. Do you automatically zcompile on login if .zwc doest exist or the respective zsh has changed ?