@ericbn 2022-01-25
> I want do be able to install/delete/update packages on my machine without worrying if it needs to dump a new compinit.

Created a `check-dumpfile` action for `zimfw` that runs the same checks as `compinit` and is intended solve the problem you describe. This was released with version [1.8.0](https://github.com/zimfw/zimfw/blob/master/CHANGELOG.md#180---2022-01-25) of zimfw.

Instead of a check during shell startup with an arbitrary schedule (e.g. 20h), this runs after files are updated by zimfw (in the build, install and update actions). Only if a change in the number of completion functions or in the Zsh version is detected, then the dumpfile is removed in order to be dumped again in the next shell startup.